use equipment sets
Instead of remembering every single slot, the original gear is now
stored in an equipment set.  This will restore _all_ slots, not just
head/feet/mainhand/offhand.

It also allows to store empty slots (like "nothing in hands" before
equipping a fishing pole) which was not possible the old way.